Remove capacity calculator classes; rename RouteFinderTest to ItineraryFinderTest
