+++ src/main/java/main/routeplanner/ItineraryFinder.java
@@ -1,5 +1,6 @@
 package main.routeplanner;
 
+import main.Route;
 import main.RouteTimetable;
 import main.Schedule;
 import main.Stop;
